.virtualenv.requirements.txt
#
# Requirement files for Python Virtual Environments
#

# Shell interface for docopt, the command-line interface description language
docopt

# YAML file parsing
PyYAML

# Access to postgresql
psycopg2

# Access to kafka
kafka-python
